Force Reload
Added a command which allows you to force the MEF and Scripts to be
reloaded. But it doesn't quite work as you need to unload the App Domain
to upload an assembly.

The compiled script dll needs to be loaded into another app domain, so
we can upload it and recompile the code. We should probably do the same
for the MEF dlls components